Update from float32 to bfloat16

+++ scripts/evaluate.py
@@ -183,7 +183,7 @@ def main():
                 num_classes=num_classes,
                 sequence_length=seq_len
             )
-            model = ModelClass(**model_cfg['model']).to(device)
+            model = ModelClass(**model_cfg['model']).to(device, dtype=torch.bfloat16)
 
             # Load Weights
             checkpoint = torch.load(ckpt_path, map_location=device)
